% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allotopeHuman.R
\name{allotope_human}
\alias{allotope_human}
\title{Function of human peptide prediction.}
\usage{
allotope_human(
  don_ag = c("DQB1_05_01", "DQB1_06_02"),
  self_ag = c("DQB1_03_01", "DQB1_03_01"),
  present_ag = c("DRB1_08_03", "DRB1_04_01"),
  seq_len = "15",
  out_name = "allotope_report.html"
)
}
\arguments{
\item{don_ag}{donor antigens.}

\item{self_ag}{self antigens.}

\item{present_ag}{presenting alleles.}

\item{seq_len}{sequence length.}

\item{out_name}{name of html report.}
}
\value{
an html report
}
\description{
Function of human peptide prediction.
}
\examples{
\dontrun{
allotope_human("/Users/newfolder")
}
}
