% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRproj.R
\name{set_proj}
\alias{set_proj}
\title{Function to set up a R project}
\usage{
set_proj(fd_name, proj_type)
}
\arguments{
\item{fd_name}{folder name.}

\item{proj_type}{project type.
working folder/project name of the project}
}
\value{
.Rproj and folder structure of the project
}
\description{
Function to set up a R project
}
\examples{
\dontrun{
set_proj("/Users/newfolder")
}
}
